Add Changes from ZSmart2.5.2b-73b10fb4236716ddcdbe4c3648da35c965ea9a8f diff with dd940d01fa2689b25ebf59c2f61cc774f6b2e398

+++ ZStack-Smart-2.6.2/Projects/zstack/ZMain/TI2530DB/OnBoard.h
@@ -238,7 +238,7 @@ typedef struct
 {
   osal_event_hdr_t hdr;
   uint8 state; // shift
-  uint8 keys;  // keys
+  uint16 keys;  // keys
 } keyChange_t;
 
 /*********************************************************************
@@ -266,7 +266,7 @@ typedef struct
   /*
    * Send "Key Pressed" message to application
    */
-  extern uint8 OnBoard_SendKeys( uint8 keys, uint8 shift );
+  extern uint8 OnBoard_SendKeys( uint16 keys, uint8 shift );
 
 /* LCD Emulation/Control Functions */
   /*
@@ -307,7 +307,7 @@ typedef struct
   /*
    * Callback routine to handle keys
    */
-  extern void OnBoard_KeyCallback ( uint8 keys, uint8 state );
+  extern void OnBoard_KeyCallback ( uint16 keys, uint8 state );
 
   /*
    * Board specific random number generator
